+++ Tests/Runtime/VespaRuntimeTests.cs
@@ -4,6 +4,5 @@ namespace LMirman.VespaIO.Tests.Runtime
 {
 	public class VespaRuntimeTests
 	{
-
 	}
 }